API improvements to fix confusing usage patterns
Most logging messages with a critical level have been bumped down to
info. This means they wont be displayed unless the logging level is
configured. This doesn't affect the CLI because the level is set to INFO.

Several routines have had references added at the top-level:

jetstream.project_init()
jetstream.load_template()
jetstream.build_workflow()
jetstream.template_environment()
jetstream.load_data_file()
jetstream.loadable_files()

Templates

Templates are now loaded from files in all cases. This change means that
the jinja2 environment needs to be configured dynamically when loading a
template (the loader needs to be configured to include the directory
where the template was located). This also means that "built-in"
templates are largely deprecated. They still exist, and the path can
be found with jetstream.built_in_templates, but the use case is small.

Runner

Minor code improvements, the runner can only be started once now.
Slurm Backend will now submit jobs with a comment that includes the
run fingerprint as a json string.

Projects

Project.init() can only be run in current directory now. This simplifies
a lot of other code, but later we might find a need for init in other
directories. Project.reload() reloads all the config data in the current
project. Two new methods are intended to aid in expanding projects:
project.pipeline and project.add_workflow. project.pipeline loads the
current workflow data from <project>/jetstream/workflow, and is meant
for inspecting or adding to the project's tasks. project.add_workflow
combines a workflow with the projects current pipeline to produce a
merged workflow. This is used by the pipelines command.

Workflows

InvalidTaskDirective was added to track errors that occur when a task is
instantiated with invalid options. Workflow.remove_task was added.
Workflow.list_tasks was added (equivilent to list(wf.tasks(objs=True)))
Workflow.compose was rewritten, this is in response to issue #11. Now,
when workflows are composed, tasks will be overwritten if their status
is "failed".