Improvements to  Task, Workflow, and WorkflowIterator
This enables aribtarary task directives, and allows a task identity
to be calculated from the directives. This allows any two tasks to
be easily compared while ignoring state.

These features were a major change to the Task API. Tasks no longer
require a unique ID to be declared in the workflow, it is calculated
from the actual content of the task itself. Flow directives, which
linked to tasks to other tasks by ID (before, after), will now link
tasks by their "name". Name does not need to be unique.

Existing workflow templates will need to be updated.