Reworked stdin directive
Originally I was launching raw commands, and it was difficult to
add bash code to the command. So i was using cmd: bash stdin: ... to
run the bash code. Now, the cmd is always executed wiht a shell anyway
so the stdin directive will be treated as a file path