Template rendering improvements
- Project config data is now added directly to the rendering data, this
means that any items in project config data (like you see with
`jetstream project config`) are available for templates to use without
accessing the project.config object. This also provides a mechanism for
overriding project data with command args. The original project config
is still available under `project.config`, this is just an addition.

- Runner will automatically complete tasks without a "cmd" directive
instead of passing them to the backend.

- Minor log/docstring fixes