Workflow split into 3 classes. This is to better control, and clarify, when the graph is being loaded for issues like #70.
Removed user/system config files from template rendering data sources. This was
something that created a lot of confusing situations when updating project
data. #67 #66 In fact, I was using it for Phoenix to store tools only, but
realized that was a mistake and have since only been using pipeline manifest.

Project history is working again #65, this may be a lot of logs if a project
has a very large config file. TODO store only a config file diff?

The arg parsers are organized better. With the changes to template data
loaded via args, the help messages should be much more clear #64

Templates can use a log() function: {{ log('hello world') }} #62

Settings, pipelines, and projects commands have been updated #61

Workflow templates can be organized as a mapping, they must include a "tasks"
key. Tasks will be loaded, anything else will be added to the workflow
as properties: Workflow.props #60

Task.identity is now calculated only from cmd and exec. This can be extended
in the future with settings, but for now I'm going to see how it works out.
The most logical addition in the future would be container ids or conda envs.
But, I want that to be added through application settings. Now, changes
to slurm args (for example) wont cause the task to be rerun. #53